requirements.txt
# Core
numpy>=1.24
pandas>=2.0
biopython>=1.81
scipy>=1.10

# UI
gradio>=4.0

# Development
pytest>=7.4
pytest-cov>=4.1

# Optimization
scikit-learn>=1.3

# ML / Embedding
torch>=2.1
transformers>=4.38
huggingface_hub>=1.3.2
